Move ecommerce-related views into ecommerce app
